Move sensitve string to different file
